python/pyproject.toml
[tool.poetry]
name = "cd74hc4067-mux-controller"
version = "0.1.0"
description = "Python library to control CD74HC4067 multiplexer via KB2040 serial interface"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
license = "MIT"
packages = [{include = "mux_controller"}]

[tool.poetry.dependencies]
python = "^3.8"
pyserial = "^3.5"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
